promise assert: Propagate assertion state to chained assertions
Since AbstractAssert.withAssertionState is package private, we now use
asInstanceOf to make an AbstractObjectAssert from the promise assert
and then use extracting to get the failure or value of the promise.
This sequence propagates the assertion state of the promise assert onto
the extracted assertion.

Also added are some InstanceOfAssertFactory members.

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>