junit5: Improve ServiceExtension
We add a public static method to create a managed ServiceConfiguration.
We use a CloseableResource to store the ServiceConfiguration in the
execution context store so that is will be properly closed when the
execution contexts completes. We also use a store keyed to the execution
context as we do in BundleContextExtension. We use the
BundleContextExtension to supply the BundleContext for the
ServiceExtension. Finally, we add support for injecting a service into
a static field.

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>