proxy: Use our class loader to create the proxy objects
We already have class loader access to all the necessary framework types
while the test bundle may not itself have access because it does not
need access.

This simplifies the API a bit, so we refactor to avoid the need to pass
a Class object into the proxy methods.

Fixes https://github.com/osgi/osgi-test/issues/229

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>