dictionary: Changes to work with assertj 3.20 and later
assertj 3.20 underwent a major change in the API related to how proxies
are handled and varargs methods. They changes the varargs methods on
the abstract super class to be final so they can be annotated with
@SafeVarargs and removed the Proxy*Assert classes used by soft
proxying to override these methods and mark them @SafeVarargs.

As a result, this broke DictionaryAssert and its soft assert badly since
DictionaryAssert extends AbstractMapAssert and override now final
methods.

So we fix the code to avoid these things allowing the code to work on
assertj 3.20 and 3.19.

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>