annotation: Substitutably import the junit4 annotation package
If both junit4 and junit5 annotation package are available, we want to
allow the junit4 bundle to import the package from the junit 5 bundle.
This allows a single version of the package to be shared for when the
tests are mixed junit4 and junit5 while preferring the junit5
annotation package over the junit4 annotation package.

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>