context: Improve error when no bundle context is available
We add a new common method to obtain the BundleContext for a test class.
If the class is not in a bundle or the bundle is not started, the
method will now throw an exception with the appropriate message.

A test class must be in a bundle loaded by an OSGi framework and that
bundle must be started (lazy activation is OK since the test runner
must have loaded the test class which would complete activation).

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>