Junit5 Extension to start an OSGi-Connect-Framework as part of the test
The current setup is rather complex and hard to understand for anyone
not familiar with OSGi+BND tooling.

This adds a JUnit5 Extension on top of the OSGi Framework connect that
drastically reduces the complexity but still enables usage of osgi-test
in a unit test environment.

Fixes https://github.com/osgi/osgi-test/issues/541