Add support for system properties in configuration annotations
This commit adds the systemProperty member to the @Property annotation. If set then the System Properties will be queried using that property name, and the value used to set the configuration. If there is no matching System Property for the supplied key then the @Property annotation value member will be used instead. If the @Property is not of type Scalar then the value of the System Property will be split using ',' as a separator.

Signed-off-by: Tim Ward <timothyjward@apache.org>