Improved support for properties in configuration annotations
This commit enhances the design of the @Property annotation to include a source. This can be used to indicate that the value of the property should be obtained from System Properties, Environment Variables, JUnit Test properties, or simply the value member of the @Property annotation (the default). If the @Property is sourced from System Properties or Environment Variables and is not of type Scalar then the discovered values will be split using ',' as a separator.

Signed-off-by: Tim Ward <timothyjward@apache.org>