Support placeholders in configuration values
Sometimes it is not sufficient to use configuration values exactly as they exist in System Properties or other sources, or several properties need to be combined to create the configuration property. This commit adds support for Value Arguments to be included in the annotations. These arguments cause the Property value to be taken as a template string suitable for use with String.format, with the supplied arguments used to populate the template.

Fixes #812

Signed-off-by: Tim Ward <timothyjward@apache.org>