Allow factory configurations to use a random name
It is possible to set the name for a factory configuration, however this is not required. Users should be permitted to omit the name member of the WithFactoryConfiguration annotation and have an auto-generated name used instead. This simplifies test setup where the configuration does not need to be injected or overridden elsewhere.

Signed-off-by: Tim Ward <timothyjward@apache.org>