Ensure that configurations can be injected without calling update
Fixes #780 - it is supposed to be possible to inject configurations that have not been updated so that tests can view the pre and post update state of the system. This was not possible (despite being hinted at in the JavaDoc) as such configurations were always 'new' and therefore were updated with an empty dictionary. This commit fixes the check for this and avoids updating configurations with no properties set if, and only if, the configuration is being injected. It also then fixes a few places where the resulting null properties caused exceptions internally in the ConfigurationExtension.

Signed-off-by: Tim Ward <timothyjward@apache.org>