Adjust compile versions for JUnit and AssertJ
The compile versions of AssertJ and JUnit gave an inconsistent base version for JUnit Jupiter, and as AssertJ only optionally requires JUnit Jupiter (it really should be another module with a mandatory dependency 🙁) you can easily end up with broken dependency graphs at runtime. This commit changes the base for AssertJ back to 3.23.x and ups JUnit to 5.8.x, which is consistent. It means we lose a couple of JUnit versions, but 5.8.x has already been released for more than two years.

Signed-off-by: Tim Ward <timothyjward@apache.org>